requirements.txt
aiogram
argon2-cffi==25.1.0
psutil
requests
qrcode
PyYAML
docker
aiohttp
aiosqlite
packaging
Jinja2
cryptography
sentry-sdk
tortoise-orm
aerich